Cleaned up my own code through making MY CODE and END MY CODE blocks.
Put all my user made functions in one spot

Created commands for starting the game, entering phase one and two, and spawning units

Begun work on the removing units function for the player to be able to change the placement however am having issues so that is commented out.

Commiting in this state because the game does not crash and things are working as intended.